LPS-53627 move common classes to the new module
