+++ modules/apps/journal/journal-content-asset-addon-entry/journal-content-asset-addon-entry-common/src/com/liferay/journal/content/asset/addon/entry/common/UserToolAssetAddonEntryTracker.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.journal.content.web.util;
+package com.liferay.journal.content.asset.addon.entry.common;
 
 import java.util.ArrayList;
 import java.util.List;
