+++ modules/apps/journal/journal-content-asset-addon-entry/journal-content-asset-addon-entry-locales/src/com/liferay/journal/content/asset/addon/entry/locales/LocalesUserToolAssetAddonEntry.java
@@ -14,7 +14,7 @@
 
 package com.liferay.journal.content.asset.addon.entry.locales;
 
-import com.liferay.journal.content.web.util.UserToolAssetAddonEntry;
+import com.liferay.journal.content.asset.addon.entry.common.UserToolAssetAddonEntry;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.servlet.taglib.ui.BaseJSPAssetAddonEntry;
