+++ modules/apps/journal/journal-content-asset-addon-entry/journal-content-asset-addon-entry-ratings/src/com/liferay/journal/content/asset/addon/entry/ratings/RatingsContentMetadataAssetAddonEntry.java
@@ -14,7 +14,7 @@
 
 package com.liferay.journal.content.asset.addon.entry.ratings;
 
-import com.liferay.journal.content.web.util.ContentMetadataAssetAddonEntry;
+import com.liferay.journal.content.asset.addon.entry.common.ContentMetadataAssetAddonEntry;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.servlet.taglib.ui.BaseJSPAssetAddonEntry;
