+++ modules/apps/journal/journal-content-asset-addon-entry/journal-content-asset-addon-entry-related-assets/src/com/liferay/journal/content/asset/addon/entry/related/assets/RelatedAssetsContentMetadataAssetAddonEntry.java
@@ -14,7 +14,7 @@
 
 package com.liferay.journal.content.asset.addon.entry.related.assets;
 
-import com.liferay.journal.content.web.util.ContentMetadataAssetAddonEntry;
+import com.liferay.journal.content.asset.addon.entry.common.ContentMetadataAssetAddonEntry;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.servlet.taglib.ui.BaseJSPAssetAddonEntry;
